chore: TokenListControlBar and useCallback memoizations
