+++ app/components/UI/Tokens/TokenListControlBar/index.ts
@@ -1,0 +1,1 @@
+export { TokenListControlBar } from './TokenListControlBar';
