+++ app/components/UI/Tokens/util/index.ts
@@ -1,2 +1,5 @@
 export { deriveBalanceFromAssetMarketDetails } from './deriveBalanceFromAssetMarketDetails';
 export { sortAssets } from './sortAssets';
+export { refreshEvmTokens } from './refreshEvmTokens';
+export { removeEvmToken } from './removeEvmToken';
+export { goToAddEvmToken } from './goToAddEvmToken';
